Extract Resource from v1 specific package

#1144
