+++ pkg/scaffold/resource/resource.go
