+++ pkg/scaffold/v1/controller/controller.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/gobuffalo/flect"
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold/v1/resource"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/resource"
 )
 
 // Controller scaffolds a Controller for a Resource
