+++ pkg/scaffold/v1/controller/controllersuitetest.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold/v1/resource"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/resource"
 )
 
 // SuiteTest scaffolds a SuiteTest
