+++ pkg/scaffold/v1/resource/register.go
@@ -20,6 +20,7 @@ import (
 	"path/filepath"
 
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/resource"
 )
 
 var _ input.File = &Register{}
@@ -29,7 +30,7 @@ type Register struct {
 	input.Input
 
 	// Resource is the resource to scaffold the types_test.go file for
-	Resource *Resource
+	Resource *resource.Resource
 }
 
 // GetInput implements input.File
