+++ pkg/scaffold/v1/webhook/add_admissionbuilder_handler.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold/v1/resource"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/resource"
 )
 
 var _ input.File = &AddAdmissionWebhookBuilderHandler{}
