+++ pkg/scaffold/v1/webhook/util.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold/v1/resource"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/resource"
 )
 
 // Use the k8s.io/api package for core resources
