+++ pkg/scaffold/v1/webhook/webhook_test.go
@@ -9,8 +9,8 @@ import (
 	. "github.com/onsi/gomega"
 	"sigs.k8s.io/kubebuilder/pkg/model"
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/resource"
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/scaffoldtest"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold/v1/resource"
 )
 
 var _ = Describe("Webhook", func() {
